+++ internal/schema/schema_test.go
@@ -15,7 +15,7 @@ func TestLoadProject(t *testing.T) {
 	recordTypePath := filepath.Join(root, "force-app/main/objects/Thing__c/recordTypes/Business.recordType-meta.xml")
 	lowercaseRecordTypePath := filepath.Join(root, "force-app/main/objects/Thing__c/recordTypes/Consumer.recordtype-meta.xml")
 	writeFile(t, objectPath, `<CustomObject xmlns="http://soap.sforce.com/2006/04/metadata"><label>Thing</label><pluralLabel>Things</pluralLabel><sharingModel>ReadWrite</sharingModel></CustomObject>`)
-	writeFile(t, fieldPath, `<CustomField xmlns="http://soap.sforce.com/2006/04/metadata"><fullName>Parent__c</fullName><label>Parent</label><type>Picklist</type><referenceTo>Thing__c</referenceTo><relationshipName>Parent__r</relationshipName><childRelationshipName>Children__r</childRelationshipName><deleteConstraint>Cascade</deleteConstraint><valueSet><valueSetDefinition><value><fullName>Hot</fullName><default>true</default><label>Hot Label</label></value><value><fullName>Cold</fullName><isActive>false</isActive></value></valueSetDefinition></valueSet></CustomField>`)
+	writeFile(t, fieldPath, `<CustomField xmlns="http://soap.sforce.com/2006/04/metadata"><fullName>Parent__c</fullName><label>Parent</label><type>Picklist</type><referenceTo>Thing__c</referenceTo><referenceTo>Account</referenceTo><relationshipName>Parent__r</relationshipName><childRelationshipName>Children__r</childRelationshipName><deleteConstraint>Cascade</deleteConstraint><valueSet><valueSetDefinition><value><fullName>Hot</fullName><default>true</default><label>Hot Label</label></value><value><fullName>Cold</fullName><isActive>false</isActive></value></valueSetDefinition></valueSet></CustomField>`)
 	writeFile(t, recordTypePath, `<RecordType xmlns="http://soap.sforce.com/2006/04/metadata"><fullName>Business</fullName><label>Business Thing</label><active>true</active><default>true</default><description>Business records</description></RecordType>`)
 	writeFile(t, lowercaseRecordTypePath, `<RecordType xmlns="http://soap.sforce.com/2006/04/metadata"><label>Consumer Thing</label><active>false</active></RecordType>`)
 
@@ -26,7 +26,7 @@ func TestLoadProject(t *testing.T) {
 	if len(s.Objects) != 1 || s.Objects[0].Name != "Thing__c" {
 		t.Fatalf("objects = %#v", s.Objects)
 	}
-	if len(s.Objects[0].Fields) != 1 || s.Objects[0].Fields[0].ReferenceTo != "Thing__c" {
+	if len(s.Objects[0].Fields) != 1 || len(s.Objects[0].Fields[0].ReferenceTo) != 2 || s.Objects[0].Fields[0].ReferenceTo[0] != "Thing__c" || s.Objects[0].Fields[0].ReferenceTo[1] != "Account" {
 		t.Fatalf("fields = %#v", s.Objects[0].Fields)
 	}
 	if got := s.Objects[0].Fields[0].DeleteConstraint; got != "Cascade" {